clear comments

+++ server/server.js
@@ -1,18 +1,15 @@
 import express from "express";
 import config from "config";
-import Sequalize from "sequelize";
 import { router } from "./routes/index.js";
 import { db } from "./models/index.js";
 
 const PORT = config.get("port");
-// const dbConnString = config.get("dbConnString");
-// const sequelize = new Sequalize(dbConnString, {
-//   logging: false
-// });
 
-db.sequelize.sync({ force: true }).then(() => {
-  console.log("Drop and re-sync db.");
-});
+if (process.env.NODE_ENV === "development") {
+  db.sequelize.sync({ force: true }).then(() => {
+    console.log("Database cleared.");
+  });
+}
 
 const app = express();
 
@@ -23,5 +20,3 @@ app.listen(PORT, () =>
 app.use(express.json());
 
 app.use("/", router);
-
-export default app;
